updates to resolve reveiw changes, including using the proper job state of , updating the documentation and moving enums to to  and adding the default reserve timeout as a constant
