+++ cli/tests/test_admin_cli.py
@@ -23,7 +23,7 @@ from http import HTTPStatus
 import pytest
 
 import testflinger_cli
-from testflinger_cli.consts import ServerRoles
+from testflinger_cli.enums import ServerRoles
 from testflinger_cli.errors import AuthorizationError
 
 from .test_cli import URL
