+++ cli/tests/test_cli_auth.py
@@ -24,7 +24,7 @@ from http import HTTPStatus
 import pytest
 
 import testflinger_cli
-from testflinger_cli.consts import ServerRoles
+from testflinger_cli.enums import ServerRoles
 from testflinger_cli.errors import NetworkError
 
 from .test_cli import URL
